[REF] tools: isolate business code
Isolate functions that are delated to business data like addresses,
partners, countries, etc.

task-4044834
